+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## WIP
+
+- Add a "min_length" parameter in the RandomizeText and RandomizeEmail converters.
+
 ## [1.0.0-beta12] - 2019-10-21
 [1.0.0-beta12]: https://git.smile.fr/dirtech/gdpr-dump/compare/1.0.0-beta11...1.0.0-beta12